+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.10
+// swift-tools-version:5.10.1
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -13,10 +13,10 @@ let package: Package = .init(
         .executable(name: "mist", targets: ["Mist"])
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-argument-parser", from: "1.6.1"),
-        // .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.57.2"),
-        .package(url: "https://github.com/SimplyDanny/SwiftLintPlugins", from: "0.61.0"),
-        .package(url: "https://github.com/jpsim/Yams", from: "6.1.0")
+        .package(url: "https://github.com/apple/swift-argument-parser", from: "1.8.2"),
+        .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.61.1"),
+        .package(url: "https://github.com/SimplyDanny/SwiftLintPlugins", from: "0.63.1"),
+        .package(url: "https://github.com/jpsim/Yams", from: "6.2.2")
     ],
     targets: [
         .executableTarget(